refactor: Moved Solver main class to its own file
