+++ src/index.ts
@@ -1,3 +1,3 @@
-export { Solver } from "./structs/2captcha"
+export { Solver } from "./structs/2captchaSolver"
 export { APIError } from "./structs/2captchaError"
 // export { Server } from "./structs/2captchaServer.js"
